packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-social-accounts"
version = "0.1.0"
description = "Account definitions, platforms, namespaces/purposes, and API keys"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["social-accounts","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-social-accounts"
Repository = "https://github.com/mxn2020/minions-social-accounts"

[tool.hatch.build.targets.wheel]
packages = ["minions_social_accounts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
